fix: refine browser config split to remove shared-utils module
- Move BrowserConfig to dedicated browser_runtime_config.py
- Inline diagnostic helpers directly in browser_diagnostics.py
- Upgrade _run_browser_diagnostics import to top-level (cycle resolved)
- Delete browser_diagnostics_utils.py (no dumping ground)
- Align _parse_remote_debugging_args to first-wins semantics
